요청제목 (#15)
* 변수 타입 다시 적음

일단 보이는대로 정리했는데... 어딘가 덜 고쳤을 수도 있음

* 타입 수정 남은거 더 함

* #14 답글달기

+++ g_board/src/main/java/com/mycompany/myapp/domain/BoardVO.java
@@ -11,7 +11,7 @@ public class BoardVO {
 	private Date bDate;
 	private Date bUpdateDate;
 	private Long bHit;
-	private int bGroup;
+	private Long bGroup;
 	private int bStep;
 	private int bIndent;
 	public Long getbNo() {
@@ -56,10 +56,10 @@ public class BoardVO {
 	public void setbHit(Long bHit) {
 		this.bHit = bHit;
 	}
-	public int getbGroup() {
+	public Long getbGroup() {
 		return bGroup;
 	}
-	public void setbGroup(int bGroup) {
+	public void setbGroup(Long bGroup) {
 		this.bGroup = bGroup;
 	}
 	public int getbStep() {
